+++ Tests/Microsoft.AppCenter.NET/Constants.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT License.
 
 using System;
-using System.IO;
 
 namespace Microsoft.AppCenter.Utils
 {
@@ -14,5 +13,7 @@ namespace Microsoft.AppCenter.Utils
         public static string AppCenterFilesDirectoryPath = "";
 
         public static string AppCenterDatabasePath = "Microsoft.AppCenter.Storage";
+
+        public static string LocalAppData => Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
     }
 }